Module actix_session::storage

Pluggable storage backends for session state.

Structs

CookieSessionStore

Use the session key, stored in the session cookie, as storage backend for the session state.

RedisSessionStore

Use Redis as session storage backend.

RedisSessionStoreBuilder

A fluent builder to construct a RedisSessionStore instance with custom configuration parameters.

SessionKey

A session key, the string stored in a client-side cookie to associate a user with its session state on the backend.

Enums

LoadError

Possible failures modes for SessionStore::load.

SaveError

Possible failures modes for SessionStore::save.

UpdateError

Possible failures modes for SessionStore::update.

Traits

SessionStore

The interface to retrieve and save the current session data from/to the chosen storage backend.