Struct actix_session::storage::RedisActorSessionStore

pub struct RedisActorSessionStore { /* private fields */ }

Use Redis as session storage backend.

use actix_web::{web, App, HttpServer, HttpResponse, Error};
use actix_session::{SessionMiddleware, storage::RedisActorSessionStore};
use actix_web::cookie::Key;

// The secret key would usually be read from a configuration file/environment variables.
fn get_secret_key() -> Key {
    // [...]
}

#[actix_web::main]
async fn main() -> std::io::Result<()> {
    let secret_key = get_secret_key();
    let redis_connection_string = "127.0.0.1:6379";
    HttpServer::new(move ||
            App::new()
            .wrap(
                SessionMiddleware::new(
                    RedisActorSessionStore::new(redis_connection_string),
                    secret_key.clone()
                )
            )
            .default_service(web::to(|| HttpResponse::Ok())))
        .bind(("127.0.0.1", 8080))?
        .run()
        .await
}

Implementation notes

RedisActorSessionStore leverages actix-redis’s RedisActor implementation - each thread worker gets its own connection to Redis.

Limitations

RedisActorSessionStore does not currently support establishing authenticated connections to Redis. Use RedisSessionStore if you need TLS support.

Implementations

impl RedisActorSessionStore

pub fn builder<S: Into<String>>(
connection_string: S
) -> RedisActorSessionStoreBuilder

A fluent API to configure RedisActorSessionStore.

It takes as input the only required input to create a new instance of RedisActorSessionStore—a connection string for Redis.

pub fn new<S: Into<String>>(connection_string: S) -> RedisActorSessionStore

Create a new instance of RedisActorSessionStore using the default configuration. It takes as input the only required input to create a new instance of RedisActorSessionStore - a connection string for Redis.