Struct actix_session::storage::RedisSessionStore

pub struct RedisSessionStore { /* private fields */ }

Use Redis as session storage backend.

use actix_web::{web, App, HttpServer, HttpResponse, Error};
use actix_session::{SessionMiddleware, storage::RedisSessionStore};
use actix_web::cookie::Key;

// The secret key would usually be read from a configuration file/environment variables.
fn get_secret_key() -> Key {
    // [...]
}

#[actix_web::main]
async fn main() -> std::io::Result<()> {
    let secret_key = get_secret_key();
    let redis_connection_string = "redis://127.0.0.1:6379";
    let store = RedisSessionStore::new(redis_connection_string).await.unwrap();

    HttpServer::new(move ||
            App::new()
            .wrap(SessionMiddleware::new(
                store.clone(),
                secret_key.clone()
            ))
            .default_service(web::to(|| HttpResponse::Ok())))
        .bind(("127.0.0.1", 8080))?
        .run()
        .await
}

TLS support

Add the redis-rs-tls-session feature flag to enable TLS support. You can then establish a TLS connection to Redis using the rediss:// URL scheme:

use actix_session::{storage::RedisSessionStore};

let redis_connection_string = "rediss://127.0.0.1:6379";
let store = RedisSessionStore::new(redis_connection_string).await.unwrap();

Implementation notes

RedisSessionStore leverages redis-rs as Redis client.

Implementations

impl RedisSessionStore

pub fn builder<S: Into<String>>(
connection_string: S
) -> RedisSessionStoreBuilder

A fluent API to configure RedisSessionStore. It takes as input the only required input to create a new instance of RedisSessionStore - a connection string for Redis.

pub async fn new<S: Into<String>>(
connection_string: S
) -> Result<RedisSessionStore, Error>

Create a new instance of RedisSessionStore using the default configuration. It takes as input the only required input to create a new instance of RedisSessionStore - a connection string for Redis.