actix_session
Module config

Configuration options to tune the behaviour of SessionMiddleware.

Structs

BrowserSession

A session lifecycle strategy where the session cookie expires when the browser’s current session ends.

PersistentSession

A session lifecycle strategy where the session cookie will be persistent.

SessionMiddlewareBuilder

A fluent, customized SessionMiddleware builder.

Enums

CookieContentSecurity

Determines how to secure the content of the session cookie.

SessionLifecycle

Determines what type of session cookie should be used and how its lifecycle should be managed.

TtlExtensionPolicy

Configuration for which events should trigger an extension of the time-to-live for your session.