Struct actix_session::Session

pub struct Session(_);

The primary interface to access and modify session state.

Session is an extractor—you can specify it as an input type for your request handlers and it will be automatically extracted from the incoming request.

use actix_session::Session;

async fn index(session: Session) -> actix_web::Result<&'static str> {
    // access session data
    if let Some(count) = session.get::<i32>("counter")? {
        session.insert("counter", count + 1)?;
    } else {
        session.insert("counter", 1)?;
    }

    Ok("Welcome!")
}

You can also retrieve a Session object from an HttpRequest or a ServiceRequest using SessionExt.

Implementations

impl Session

pub fn get<T: DeserializeOwned>(&self, key: &str) -> Result<Option<T>, Error>

Get a value from the session.

It returns an error if it fails to deserialize as T the JSON value associated with key.

pub fn entries(&self) -> Ref<'_, HashMap<String, String>>

Get all raw key-value data from the session.

Note that values are JSON encoded.

pub fn status(&self) -> SessionStatus

Returns session status.

pub fn insert(
&self,
key: impl Into<String>,
value: impl Serialize
) -> Result<(), Error>

Inserts a key-value pair into the session.

Any serializable value can be used and will be encoded as JSON in session data, hence why only a reference to the value is taken.

It returns an error if it fails to serialize value to JSON.

pub fn remove(&self, key: &str) -> Option<String>

Remove value from the session.

If present, the JSON encoded value is returned.

pub fn remove_as<T: DeserializeOwned>(
&self,
key: &str
) -> Option<Result<T, String>>

Remove value from the session and deserialize.

Returns None if key was not present in session. Returns T if deserialization succeeds, otherwise returns un-deserialized JSON string.

pub fn clear(&self)

Clear the session.

pub fn purge(&self)

Removes session both client and server side.

pub fn renew(&self)

Renews the session key, assigning existing session state to new key.